requirements.txt
numpy
pandas
pytorch
scikit-learn
nltk
